[Web/Backend][Test] #45 파일 구조 변경, utils, service 폴더 정리
