+++ web/Backend/express/service/clovaSpeech.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
-const {getSecret} = require("../gcs");
-
+const {getSecret} = require("./gcsService");
+const {clovaTimestamping} = require("../utils/clovaTimestamping");
 
 const CLOVA_API_URL = 'https://clovaspeech-gw.ncloud.com/external/v1/11266/6a0ff352edd40a4698cf042f549658cccd39feafa0875763efa41eaa99a79a72/recognizer/url';
 
@@ -32,4 +32,17 @@ async function requestClova(audiourl) {
     return response.data;
 }
 
-module.exports.requestClova = requestClova;
+
+async function clovaSTT(audiourl) {
+    try {
+        const sttresult = await requestClova(audiourl); //stt 요청
+        const result = sttresult.segments.map(clovaTimestamping);//타임스탬프 분리
+        console.log(`✅ 타임스탬프 처리 완료`);
+        return result;
+    } catch (error) {
+        console.error('❌ 에러 발생:', error.response?.data || error.message);
+    }
+}
+
+
+module.exports.clovaSTT = clovaSTT;
