+++ web/Backend/express/utils/gcsSttFormatter.js
